+++ src/recorder/cache/VideoCache.js
@@ -73,7 +73,7 @@ async function storeBlob(getStore, blob, key) {
 }
 
 /**
- * @param {import('./DocumentStore').TakeId} takeId
+ * @param {import('../../stores/DocumentStore').TakeId} takeId
  * @param {Blob} blob
  */
 export async function cacheVideoBlob(takeId, blob) {
@@ -89,7 +89,7 @@ export async function cacheVideoBlob(takeId, blob) {
 }
 
 /**
- * @param {import('./DocumentStore').TakeId} takeId
+ * @param {import('../../stores/DocumentStore').TakeId} takeId
  */
 export async function getVideoBlob(takeId) {
   return new Promise((resolve, reject) => {
@@ -103,7 +103,7 @@ export async function getVideoBlob(takeId) {
 }
 
 /**
- * @param {import('./DocumentStore').TakeId} takeId
+ * @param {import('../../stores/DocumentStore').TakeId} takeId
  */
 export async function deleteVideoBlob(takeId) {
   return new Promise((resolve, reject) => {